feat(database): swap migration runner to goose (ADR-0006, Phase 5b-2)
Point db.migrate() at goose over the embedded baseline and delete the
2.2k-line homegrown index+1 runner. The connection setup (PRAGMAs, WAL,
pool) in database.go is untouched; only how the schema is created and
tracked changed.

- goose.go: the new engine — //go:embed migrations/*.sql + gooseProvider,
  migrate() (goose Up), and MigrationStatus()/SchemaVersion() rewired to
  goose Status/GetDBVersion (both were test-only). The schema_migrations
  bookkeeping table gives way to goose_db_version (already excluded from
  the schema snapshot).
- migrations.go (getMigrationDefs/getMigrations/runMigration/
  getCurrentVersion + the Migration/migrationDef types) DELETED.
- seed.go: seedDefaultProfile moved out of the old engine, unchanged.

A real gap the schema-only snapshot could not catch surfaced under test
and is fixed: the legacy migrations INSERTed a default `clients` row
(profiles.client_id REFERENCES clients(id) defaults to 'default'), which
a sqlite_master dump does not capture. That seed is now carried in the
baseline's Up section, so seedDefaultProfile's FK is satisfiable on a
fresh DB. The full ./internal/database suite confirms no other seed data
was lost.

The one-time generator that derived the baseline from the legacy runner
is removed with that runner (it depended on it); TestGooseBaselineRepro-
ducesSchema (baseline -> goose -> schema == golden) and TestSchemaSnapshot
remain as the equivalence gates. ExportMigrationsCount now counts the
embedded .sql baselines (= 1).

Gates: full ./internal/database suite race-clean; snapshot byte-identical;
golangci 0; binary builds. STRICT-table conversion is 5b-3.
